add: user defined uuids (#438)
This pull request enhances the Rawstor CLI by providing more granular control over object creation and improving data type robustness. By enabling user-defined UUIDs and direct target specification, the CLI becomes more versatile for complex deployment scenarios. Additionally, the migration to 64-bit integers for size calculations ensures that the system can handle modern storage capacities accurately across different architectures.

### Highlights

* **User-Defined UUIDs**: Added support for providing explicit UUIDs during object creation via the new `-u` command-line option.
* **CLI Flexibility**: Introduced the `-t` (target) option, allowing users to create objects using pre-formatted target strings as an alternative to specifying locations.
* **64-bit Size Support**: Updated internal size handling to use `uint64_t` across the CLI and core libraries, ensuring support for larger object sizes and preventing overflow on 32-bit systems.
* **Expanded Unit Support**: Extended unit parsing and formatting to include Peta (P) and Exa (E) prefixes.

+++ cli/testio.c
@@ -21,7 +21,7 @@ typedef struct {
 } Worker;
 
 static Worker* worker_create(
-    unsigned int index, size_t block_size, unsigned int* counter,
+    unsigned int index, uint64_t block_size, unsigned int* counter,
     unsigned int niterations
 ) {
     Worker* worker = malloc(sizeof(Worker));
@@ -256,7 +256,7 @@ static int srcv_data_sent(
 }
 
 int rawstor_cli_testio(
-    unsigned int queue_size, const char* target, size_t block_size,
+    unsigned int queue_size, const char* target, uint64_t block_size,
     unsigned int count, unsigned int io_depth, int vector_mode
 ) {
     int res;
